requirements.txt
sentence-transformers==2.7.0
requests==2.31.0
python-dotenv==1.0.1
streamlit==1.33.0
pymupdf==1.24.1
groq==0.9.0
rich==13.7.1
